requirements.txt
discord.py
aiohttp
urlextract